+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Other
+
+- Package and extension scaffolds emit least-privilege manifest policy metadata ([pm-bav0](https://github.com/unbraind/pm-cli/blob/main/.agents/pm/chores/pm-bav0.toon))
+
 ## 2026.6.21 - 2026-06-21
 
 ### Added